[Ghidra] Run with options loaded with sensible defaults
Was trailofbits/ManticoreUI-Ghidra#2

PR Description:
As per trailofbits/ManticoreUI#29

Adds a small UI change that allows user to quickly alter important manticore options.

Options presented are based on the run_dialog in the binja plugin. Validation is handled by manticore itself, where if an option is invalid the stack trace is printed to the MUI Log.

Co-authored-by: Eric Kilmer <eric.d.kilmer@gmail.com>